+++ src/main/java/com/kyg/intro/data/repository/ScheduleRepository.java
@@ -3,5 +3,5 @@ package com.kyg.intro.data.repository;
 import com.kyg.intro.data.entity.Schedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScheduleRepository extends JpaRepository<Schedule, Integer>{
+public interface ScheduleRepository extends JpaRepository<Schedule, Long>{
 }
